Move SubNetwork functionality to Node
